s/t.integer/t.bigint/g

+++ db/schema.rb
@@ -2,8 +2,8 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# This file is the source Rails uses to define your schema when running `rails
-# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
 # be faster and is potentially less error prone than running all of your
 # migrations from scratch. Old migrations may fail to apply correctly if those
 # migrations use external dependencies or application code.
@@ -21,7 +21,7 @@ ActiveRecord::Schema.define(version: 2020_11_15_120803) do
   create_table "puppet_nodes", force: :cascade do |t|
     t.string "fqdn"
     t.string "role"
-    t.integer "puppet_environment_id", null: false
+    t.bigint "puppet_environment_id", null: false
     t.string "zone"
     t.string "os_family"
     t.string "os_lsbdistcodename"
@@ -43,7 +43,7 @@ ActiveRecord::Schema.define(version: 2020_11_15_120803) do
     t.string "last_name"
     t.string "email"
     t.string "password_digest"
-    t.integer "role_id", null: false
+    t.bigint "role_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["email"], name: "index_users_on_email", unique: true
